requirements.txt
logzero==1.5.0
influxdb==5.3.0
speedtest-cli==2.1.2